requirements.txt
python-docx
streamlit
PyPDF2
requests
python-dotenv
google.generativeai
